+++ qq-bridge/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @sfmc-bds/qq-bridge
 
+## 0.1.1-beta.4
+
+### Patch Changes
+
+- c1de2a8: none
+- f8cbe3b: none
+- b6f8adc: none
+- Updated dependencies [c1de2a8]
+- Updated dependencies [f8cbe3b]
+- Updated dependencies [b6f8adc]
+  - @sfmc-bds/sdk@0.2.0-beta.6
+
+## 0.1.1-beta.3
+
+### Patch Changes
+
+- none
+- Updated dependencies
+  - @sfmc-bds/sdk@0.2.0-beta.5
+
+## 0.1.1-beta.2
+
+### Patch Changes
+
+- none
+- Updated dependencies
+  - @sfmc-bds/sdk@0.2.0-beta.4
+
+## 0.1.1-beta.1
+
+### Patch Changes
+
+- none
+- Updated dependencies
+  - @sfmc-bds/sdk@0.2.0-beta.3
+
 ## 0.1.1-beta.0
 
 ### Patch Changes